+++ src/dipdup/models.py
@@ -10,7 +10,8 @@ from pydantic.dataclasses import dataclass
 from pydantic.error_wrappers import ValidationError
 from tortoise import Model, fields
 
-from dipdup.exceptions import ConfigurationError, DipDupException, InvalidDataError, ReindexingReason
+from dipdup.enums import IndexStatus, IndexType
+from dipdup.exceptions import ConfigurationError, InvalidDataError, ReindexingReason
 
 ParameterType = TypeVar('ParameterType', bound=BaseModel)
 StorageType = TypeVar('StorageType', bound=BaseModel)
@@ -19,19 +20,6 @@ ValueType = TypeVar('ValueType', bound=BaseModel)
 
 
 _logger = logging.getLogger('dipdup.models')
-
-
-class IndexType(Enum):
-    operation = 'operation'
-    big_map = 'big_map'
-
-
-class IndexStatus(Enum):
-    NEW = 'NEW'
-    SYNCING = 'SYNCING'
-    REALTIME = 'REALTIME'
-    ROLLBACK = 'ROLLBACK'
-    ONESHOT = 'ONESHOT'
 
 
 @dataclass
@@ -299,26 +287,17 @@ class Index(Model):
     template_values = fields.JSONField(null=True)
 
     level = fields.IntField(default=0)
-    head = fields.ForeignKeyField('int_models.Head', related_name='indexes', null=True)
 
     created_at = fields.DatetimeField(auto_now_add=True)
     updated_at = fields.DatetimeField(auto_now=True)
 
     async def update_status(
         self,
-        status: IndexStatus,
+        status: Optional[IndexStatus] = None,
         level: Optional[int] = None,
-        head: Optional['Head'] = None,
     ) -> None:
-        if level:
-            if level < self.level:
-                raise DipDupException('Index level is higher than desired level')
-            self.level = level  # type: ignore
-
-        if head:
-            self.head = head
-
-        self.status = status
+        self.status = status or self.status
+        self.level = level or self.level  # type: ignore
         await self.save()
 
     class Meta: